requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
auto-py-to-exe==2.8.0
Automat==20.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Click==7.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
cssselect==1.1.0
decorator==4.4.2
distlib==0.3.1
Django==3.1.5
Eel==0.12.4
filelock==3.0.12
Flask==1.0.2
Flask-SQLAlchemy==2.4.4
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
hyperlink==21.0.0
idna==2.10
image==1.5.33
incremental==21.3.0
invoke==1.5.0
ipython==7.21.0
ipython-beautifulsoup==0.3
ipython-genutils==0.2.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.10.3
jmespath==0.10.0
key==0.4
keyboard==0.13.5
lxml==4.6.2
MarkupSafe==1.1.1
MouseInfo==0.1.3
numpy==1.19.5
opencv-python==4.5.1.48
pandas==1.2.1
parsel==1.6.0
parso==0.8.1
pefile==2019.4.18
pickleshare==0.7.5
pillow==8.1.0
prompt-toolkit==3.0.17
Protego==0.1.16
py2exe==0.10.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
pycparser==2.20
PyDispatcher==2.0.5
PyGetWindow==0.0.9
Pygments==2.8.1
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyOpenSSL==20.0.1
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.2.6
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.5
pywin32==300
pywin32-ctypes==0.2.0
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.2
SQLAlchemy==1.3.22
sqlparse==0.4.1
Tcl==0.2
tqdm==4.56.0
traitlets==5.0.5
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urljoin==1.0.0
urllib3==1.26.3
virtualenv==20.4.0
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==0.16.0
whichcraft==0.6.1
WTForms==2.3.3
xmljson==0.2.1
zope.event==4.5.0
zope.interface==5.2.0
